fix: SimplePathPatternParser 코드 관련 수정, DefaultConversionService 코드 관련 수정

+++ src/test/java/com/milktea/myspring/boot/web/utils/converters/DefaultConversionServiceTest.java
@@ -27,7 +27,7 @@ public class DefaultConversionServiceTest {
         String source = "2";
 
         //when
-        Object result = conversionService.convert(source, Integer.class);
+        Object result = conversionService.convert(source, Long.class);
 
         //then
         Assertions.assertTrue(result instanceof Long);
